Gave the ships different values

+++ InnerFear/Classes/ShipFactory.h
@@ -12,15 +12,15 @@
 
 
 
-@property (nonatomic, assign) int unitSize;
+@property (nonatomic, assign) double unitSize;
 @property (nonatomic, assign) double timeForUnit;
 @property (nonatomic, assign) double currentTime;
 @property (nonatomic, assign) double accumulatedUnits;
 @property (nonatomic, retain) NSMutableArray* shiporder;
 
 
-- (id) initWithTimeForUnit: (double) timeForUnit unitSize: (int) unitSize;
-- (int) getUnit: (double) timePassed;
+- (id) initWithTimeForUnit: (double) timeForUnit unitSize: (double) unitSize;
+- (double) getUnit: (double) timePassed;
 - (NSArray*) getBuildUnits: (double) timePassed;
 
 @end
